+++ app/src/main/java/com/yogigupta1206/invoicereceiptmaker/domain/repository/TnCRepository.kt
@@ -7,7 +7,7 @@ interface TnCRepository {
 
     fun getAllTnc(): Flow<List<TnC>>
 
-    suspend fun getTncById(id: Int): TnC
+    suspend fun getTncById(id: Long): TnC
 
     suspend fun insertTnc(tnc: TnC)
